refactor(solver): split generic call finalization
